[scsynth:portaudio] move helper functions
